pyproject.toml
[project]
name = "manimgeneration"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruthvik Konduru",email = "114494803+BRKon123@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain (>=0.3.17,<0.4.0)",
    "langgraph (>=0.2.68,<0.3.0)",
    "langchain-anthropic (>=0.3.5,<0.4.0)",
    "langsmith (>=0.3.3,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "manim (==0.18.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
